requirements.txt
python-telegram-bot==21.7
flask==3.0.3
python-dotenv==1.0.1
pyngrok==7.2.2
